+++ src/ts/client/view/state/gamemodeListOfPlayersPlayed.ts
@@ -2,7 +2,7 @@ import { IGameModeItemWithoutCollections } from "../../../type/list/IGameModeIte
 import { IGameSystemInfoWithoutCollections } from "../../../type/list/IGameSystemInfo";
 import { IRunner } from "../../../type/record/IRunner";
 import { choiceString, selectAppropriateDescription } from "../../../utility/aboutLang";
-import { convertNumberIntoDateString } from "../../../utility/timeUtility";
+import { convertNumberIntoDateString } from "../../../utility/timeExpressionUtility";
 import { elementWithoutEscaping } from "../../../utility/ViewUtility";
 import { IAppUsedToReadAndChangePage } from "../../interface/AppInterfaces";
 import { generateIcooonHTML, writeElement } from "../../utility/aboutElement";
