+++ src/ts/utility/timeExpressionUtility.ts
